Formating revision.

+++ module-shared/test/openllet/aterm/test/Test1.java
@@ -5,14 +5,14 @@
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
  *
- *     * Redistributions of source code must retain the above copyright
- *       notice, this list of conditions and the following disclaimer.
- *     * Redistributions in binary form must reproduce the above copyright
- *       notice, this list of conditions and the following disclaimer in the
- *       documentation and/or other materials provided with the distribution.
- *     * Neither the name of the University of California, Berkeley nor the
- *       names of its contributors may be used to endorse or promote products
- *       derived from this software without specific prior written permission.
+ * * Redistributions of source code must retain the above copyright
+ * notice, this list of conditions and the following disclaimer.
+ * * Redistributions in binary form must reproduce the above copyright
+ * notice, this list of conditions and the following disclaimer in the
+ * documentation and/or other materials provided with the distribution.
+ * * Neither the name of the University of California, Berkeley nor the
+ * names of its contributors may be used to endorse or promote products
+ * derived from this software without specific prior written permission.
  *
  * THIS SOFTWARE IS PROVIDED BY THE REGENTS AND CONTRIBUTORS ``AS IS'' AND ANY
  * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
@@ -35,10 +35,6 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
-
-import org.junit.Before;
-import org.junit.Test;
-
 import openllet.aterm.AFun;
 import openllet.aterm.ATerm;
 import openllet.aterm.ATermAppl;
@@ -50,6 +46,8 @@ import openllet.aterm.ATermReal;
 import openllet.aterm.ParseError;
 import openllet.aterm.pure.PureFactory;
 import openllet.atom.OpenError;
+import org.junit.Before;
+import org.junit.Test;
 
 public class Test1
 {
@@ -75,8 +73,7 @@ public class Test1
 
 	public void assertTrue(final boolean condition) // TODO : change to junit.
 	{
-		if (!condition)
-			throw new AssertionError("assertion failed.");
+		if (!condition) throw new AssertionError("assertion failed.");
 	}
 
 	private static void test(final boolean cond, final String id)
@@ -218,8 +215,7 @@ public class Test1
 		}
 		catch (final ParseError e)
 		{
-			if (!e.getMessage().startsWith("Unterminated quoted function symbol"))
-				throw e;
+			if (!e.getMessage().startsWith("Unterminated quoted function symbol")) throw e;
 		}
 	}
 
@@ -230,8 +226,7 @@ public class Test1
 		{
 			final File file = new File("testAtermFileParser.txt");
 
-			if (file.exists())
-				file.delete();
+			if (file.exists()) file.delete();
 
 			try (FileOutputStream output = new FileOutputStream(file))
 			{
@@ -244,8 +239,7 @@ public class Test1
 				System.out.println("result = " + result);
 			}
 
-			if (file.exists())
-				file.delete();
+			if (file.exists()) file.delete();
 		}
 		catch (final FileNotFoundException e1)
 		{
@@ -423,7 +417,7 @@ public class Test1
 		System.out.println("result = " + result);
 		test(result != null && result.size()==1 &&
 		   result.get(0).equals(T[0]), "match-8");
-
+		
 		result = T[0].match("<f(1,2,<int>)>");
 		System.out.println("result = " + result);
 		test(result != null && result.size() == 2, "match-9a");
@@ -505,31 +499,31 @@ public class Test1
 		/*
 		list.add(new Integer(1));
 		test(_factory.make("<int>", list).equals(T[0]), "make-1");
-
+		
 		list.clear(); list.add(T[3]);
 		test(_factory.make("<term>", list).equals(T[3]), "make-2");
-
+		
 		list.clear(); list.add( "b");
 		test(_factory.make("<appl>", list).equals(T[4]), "make-3");
-
+		
 		list.clear(); list.add(new Double(3.14));
 		test(_factory.make("<real>", list).equals(
 		_factory.makeReal(3.14)), "make-4");
-
+		
 		list.clear(); list.add(_factory.makeAppl(
 		_factory.makeAFun("real",0,false)));
 		test(_factory.make("<placeholder>", list).equals(
 		_factory.parse("<real>")), "make-5");
-
+		
 		list.clear(); list.add(T[7]);
 		test(_factory.make("[<list>]", list).equals(T[7]), "make-6");
-
+		
 		list.clear();
 		list.add(T[3]);
 		list.add("b");
 		list.add(_factory.makeList(T[5], _factory.makeList()));
 		test(_factory.make("f(<term>,<appl>,<list>)", list).equals(T[6]), "make-7");
-
+		
 		list.clear();
 		list.add("f");
 		list.add(new Integer(2));
@@ -553,8 +547,7 @@ public class Test1
 		//		System.out.println("array1");
 		for (int i = 0; i < size; i++)
 		{
-			if (i % 100 == 0)
-				System.out.print(i + "  ");
+			if (i % 100 == 0) System.out.print(i + "  ");
 
 			final int idx = i % 10;
 			array1[idx] = _factory.makeAppl(a);
@@ -565,8 +558,7 @@ public class Test1
 		//		System.out.println("\narray2");
 		for (int i = 0; i < size; i++)
 		{
-			if (i % 100 == 0)
-				System.out.print(i + "  ");
+			if (i % 100 == 0) System.out.print(i + "  ");
 
 			final int idx = i % 10;
 			array2[idx] = _factory.makeAppl(a);
@@ -577,8 +569,7 @@ public class Test1
 		//		System.out.println("\ntest");
 		for (int i = 0; i < size; i++)
 		{
-			if (i % 500 == 0)
-				System.out.print(i + "  ");
+			if (i % 500 == 0) System.out.print(i + "  ");
 
 			final int idx = i % 10;
 			if (array1[idx] != array2[idx])
